refactor(cli): 拆 17 个 handle* 到 commands/<name>.ts，index.ts 改为 dispatcher
